Add PreShopCreationItemDisplayNameEvent class
This commit introduces a new class, PreShopCreationItemDisplayNameEvent, which will handle the display name of the items during shop creation. Accompanying changes have also been made to the ItemChecker class to utilize this new event and ensure proper handling of item display names, improving the functionality of the shop creation process.